### 2024-07-03 #6905 @bjmt
Ah sorry, I assume it is checking the branch I specified for the paper.md file. I can add the license there as well.
