### 2024-07-19 #7012 @editorialbot
[REVIEW]: universalmotif: An R package for biological motif analysis
**Submitting author:** @bjmt (<a href="http://orcid.org/0000-0002-7441-2951">Benjamin Jean-Marie Tremblay</a>)
**Repository:** https://github.com/bjmt/universalmotif
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.23.2
