### 2024-07-23 bjmt/universalmotif#27 @chrisamiller
[JOSS REVIEW] No community/contribution guidelines
This project lacks community guidelines, as described in the JOSS review checklist.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

A CONTRIBUTING.md file would solve this easily enough
