### 2024-07-23 #7012 @chrisamiller
Overall, this is a well-developed package, as evidenced by it's useful and clear functions, extensive documentation, and by it's inclusion in the R/Bioconductor repository. The paper is quite concise and while it could be expanded with some examples or images, the vignettes for the tool already cover that well. As is, it adequately describes the tool and it's primary use cases.

I was able to install the tool and run selected examples from the vignettes, as well as import some of my own data from HOMER and manipulate and plot it easily. The polish is impressive, and this should be a useful tool for people doing motif analyses.

My only issue is that the project lacks Community guidelines: https://github.com/bjmt/universalmotif/issues/27
