### 2024-07-24 2df091d
Add a CONTRIBUTING.md (#27)
The CONTRIBUTING.md file is based on the following template:
https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md
