### 2024-07-24 bjmt/universalmotif#27 @bjmt
A CONTRIBUTING.md file has been created with guidance on opening issues and pull requests, as well as directing users needing help to the Bioconductor support site. Please let me know if you find this suitable or have additional advice.
