### 2024-07-24 #7012 @bjmt
Dear Chris,

Thank you very much for the review. I appreciate the kind words. I have added a basic CONTRIBUTING.md file to the repository and pushed the changes to Bioconductor. Please let me know if you have any further suggestions.

Thanks,

Benjamin
