### 2024-07-24 #6905 @spholmes
Hi @ajank ,
I know it's summer but maybe you can have a few extra cycles...
I am still looking for a second reviewer for this package. It is a mature Bioconductor package so maybe not too much work to review quickly, if you could help me out here, I'd be very grateful...
Susan
