### 2024-07-25 #7012 @spholmes
Hi @ajank : you can generate the checklist with
the call to `@editorialbot generate my checklist`
Thanks for your help, susan
