### 2024-08-08 df396a9
fix Rcpp indexing warnings
