+++ src/universalmotif-class.cpp
@@ -82,7 +82,7 @@ Rcpp::StringVector universalmotif_type(Rcpp::NumericMatrix &m_motif,
   Rcpp::LogicalVector mat_099_101_check = (motif_colsums > 0.99) & (motif_colsums < 1.01);
   Rcpp::LogicalVector mat_pos_check = m_motif >= 0;
 
-  if (Rcpp::StringVector::is_na(type[0]) || type.size() == 0) {
+  if (type.size() == 0 || Rcpp::StringVector::is_na(type[0])) {
     if (Rcpp::is_true(Rcpp::all(mat_1_0_check)))
       type = Rcpp::StringVector::create("PCM");
     else if (Rcpp::is_true(Rcpp::all(mat_099_101_check))
@@ -107,7 +107,7 @@ Rcpp::NumericVector universalmotif_nsites(Rcpp::NumericVector nsites,
     const Rcpp::StringVector &type, Rcpp::NumericMatrix &m_motif,
     const Rcpp::NumericVector &motif_colsums) {
 
-  if (Rcpp::NumericVector::is_na(nsites[0]) || nsites.size() == 0) {
+  if (nsites.size() == 0 || Rcpp::NumericVector::is_na(nsites[0])) {
 
     if (type[0] == "PCM") nsites[0] = Rcpp::sum(m_motif(Rcpp::_, 0));
     else nsites = Rcpp::NumericVector::create();
@@ -147,7 +147,7 @@ Rcpp::NumericVector universalmotif_icscore(Rcpp::NumericVector icscore,
   R_xlen_t alph_len = m_motif.nrow();
   Rcpp::IntegerVector bkg_i = Rcpp::seq_len(alph_len) - 1;
 
-  if (Rcpp::NumericVector::is_na(icscore[0]) || icscore.size() == 0) {
+  if (icscore.size() == 0 || Rcpp::NumericVector::is_na(icscore[0])) {
     double tmp_nsites = 0;
     if (nsites.size() != 0) tmp_nsites = nsites[0];
     Rcpp::NumericVector icscore_tmp(m_motif.ncol());
@@ -157,7 +157,7 @@ Rcpp::NumericVector universalmotif_icscore(Rcpp::NumericVector icscore,
           Rcpp::as<std::vector<double>>(bkg[bkg_i]), Rcpp::as<std::string>(type[0]),
           pseudocount, tmp_nsites);
     }
-    icscore[0] = Rcpp::sum(icscore_tmp);
+    icscore = Rcpp::NumericVector::create(Rcpp::sum(icscore_tmp));
   }
 
   return icscore;
@@ -191,7 +191,7 @@ Rcpp::NumericVector universalmotif_bkg(Rcpp::NumericVector bkg,
   R_xlen_t alph_len = m_motif.nrow();
   R_xlen_t bkg_len = bkg.size();
 
-  if (Rcpp::NumericVector::is_na(bkg[0]) || bkg_len == 0) {
+  if (bkg_len == 0 || Rcpp::NumericVector::is_na(bkg[0])) {
 
     bkg = Rcpp::rep(1.0 / m_motif.nrow(), m_motif.nrow());
     bkg.attr("names") = Rcpp::rownames(m_motif);
@@ -580,7 +580,7 @@ Rcpp::S4 universalmotif_cpp(
   Rcpp::NumericMatrix m_motif = Rcpp::clone(motif);
   Rcpp::NumericVector motif_colsums = Rcpp::colSums(m_motif);
   // sometimes nsites can slip through (?) as nan (not R_NaN)
-  if (std::isnan(nsites[0])) nsites = Rcpp::NumericVector::create();
[…]
