### 2024-08-15 #7012 @ajank
This manuscript describes a well implemented and extensively documented R/Bioconductor package for handling biological sequence motifs. The vignettes provide detailed examples of typical use cases. It is worth emphasizing that the package functions, along with their options, are well-thought, logically named, and highly universal, allowing for a broad applicability. The package is well integrated in the Bioconductor ecosystem, allowing for interoperability with other packages, both focusing on handling motifs (e.g. TFBSTools and motifStack) and providing complementary features (e.g. ggbio and EnrichedHeatmap).

The paper describes the statement of need for this tool, and briefly describes its most important possible applications, giving the names of the relevant functions.

I have only two issues, both regarding the manuscript text:
1. State of the field: interoperability with TFBSTools and motifStack is mentioned (lines 36-40), but it would be worth to clearly state in 2-3 sentences the advantages over these other packages.
2. Widespread use is emphasized (lines 56-60), but the long list of references would be more informative if the examples were grouped by usage pattern (motif enrichment assessment, clustering similar motifs, ...). Perhaps the citations where the package potential was used only marginally could be omitted.
