### 2024-08-18 8b25325
literature examples grouped by usage pattern
