+++ paper.bib
@@ -239,26 +239,6 @@
 	keywords = {Algorithms, Animals, Biopolymers, Humans, Models, Theoretical, Sequence Analysis},
 }
 
-@article{gibson24,
-	title = {Protein-intrinsic properties and context-dependent effects regulate pioneer factor binding and function},
-	volume = {31},
-	rights = {2024 The Author(s), under exclusive licence to Springer Nature America, Inc.},
-	issn = {1545-9985},
-	url = {https://www.nature.com/articles/s41594-024-01231-8},
-	doi = {10.1038/s41594-024-01231-8},
-	abstract = {Chromatin is a barrier to the binding of many transcription factors. By contrast, pioneer factors access nucleosomal targets and promote chromatin opening. Despite binding to target motifs in closed chromatin, many pioneer factors display cell-type-specific binding and activity. The mechanisms governing pioneer factor occupancy and the relationship between chromatin occupancy and opening remain unclear. We studied three Drosophila transcription factors with distinct {DNA}-binding domains and biological functions: Zelda, Grainy head and Twist. We demonstrated that the level of chromatin occupancy is a key determinant of pioneering activity. Multiple factors regulate occupancy, including motif content, local chromatin and protein concentration. Regions outside the {DNA}-binding domain are required for binding and chromatin opening. Our results show that pioneering activity is not a binary feature intrinsic to a protein but occurs on a spectrum and is regulated by a variety of protein-intrinsic and cell-type-specific features.},
-	pages = {548--558},
-	number = {3},
-	journaltitle = {Nat Struct Mol Biol},
-	author = {Gibson, Tyler J. and Larson, Elizabeth D. and Harrison, Melissa M.},
-	urldate = {2024-06-17},
-	date = {2024-03},
-	langid = {english},
-	note = {Publisher: Nature Publishing Group},
-	keywords = {Chromatin, Molecular biology, Reprogramming, Transcription factors},
-	file = {Full Text PDF:/Users/ben/Zotero/storage/G5I2JGAV/Gibson et al. - 2024 - Protein-intrinsic properties and context-dependent.pdf:application/pdf},
-}
-
 @article{hawkins24,
 	title = {{ePRINT}: exonuclease assisted mapping of protein-{RNA} interactions},
 	volume = {25},
