### 2024-08-18 #7012 @bjmt
Dear Aleksander,

Thank you very much for the positive review, your kind words mean a lot to me. To address your concerns:

1. I have added a few sentences to compare the universalmotif package with other available R/Bioconductor motif packages, first in terms of the class conversion utilities unique to universalmotif, and second in terms of the extensiveness of the functionality available (bjmt/universalmotif@57ca7324803669ef0a26815941a57891e20ddf81). However, I hope you can allow me to avoid overly explicit mentions of advantages of the universalmotif package over others, as I have always wished for this package to be seen to work alongside existing packages, and not be a complete replacement.
2. The final paper examples have been reorganized by which methods are stated were used (bjmt/universalmotif@8b253259125036fe0e8e14e85b4cc53a3a9a2e22). One paper which did not specify how the universalmotif package was used exactly was removed.

Please let me know if you have further suggestions.

Sincerely,

Benjamin
