### 2024-08-19 #7012 @bjmt
The latest version (1.23.2) has been archived on [Zenodo](https://zenodo.org/records/13341141) (10.5281/zenodo.13341141) using the same license and author information.
