### 2024-08-19 #7012 @Kevin-Mattheus-Moerman
- [ ] Please edit the ZENODO archive such that the version includes the `v` in `v1.23.2`, i.e. it should fully match your GitHub released version tag.
