### 2024-08-20 29db0f7
add citation to readme
