+++ CITATION.cff
@@ -1,0 +1,26 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Tremblay
+  given-names: Benjamin Jean-Marie
+  orcid: "https://orcid.org/0000-0002-7441-2951"
+doi: 10.5281/zenodo.13341141
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Tremblay
+    given-names: Benjamin Jean-Marie
+    orcid: "https://orcid.org/0000-0002-7441-2951"
+  date-published: 2024-08-19
+  doi: 10.21105/joss.07012
+  issn: 2475-9066
+  issue: 100
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7012
+  title: "universalmotif: An R package for biological motif analysis"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07012"
+  volume: 9
+title: "universalmotif: An R package for biological motif analysis"
